requirements.txt
flask
flask-cors
openai
anthropic
google-genai
python-dotenv
requests
gunicorn
PyPDF2
Pillow
